# packages/api-client/openapi/v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Shellwright control plane",
    "description": "Organisations, apps, and immutable configuration versions.\n\nErrors are RFC 9457 problem documents. The `type` URI and the `code` extension are stable; the `title` and `detail` are for people and may be reworded.",
    "version": "v1"
  },
  "paths": {
...
    "/v1/apps/{appId}/config": {
      "get": {
        "tags": ["Configuration"],
        "summary": "Read the current resolved configuration.",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionResponse"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified"
          }
        }
      },
      "post": {
        "tags": ["Configuration"],
        "summary": "Save a new configuration version.",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/apps/{appId}/config/validate": {
      "post": {
        "tags": ["Configuration"],
        "summary": "Validate without saving.",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/apps/{appId}/config/versions": {
      "get": {
        "tags": ["Configuration"],
        "summary": "List configuration versions, newest first.",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "string"
                }
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOfVersionSummary"
                }
              }
            }
          }
        }
      }
    },
    "/v1/apps/{appId}/config/versions/{versionId}": {
      "get": {
        "tags": ["Configuration"],
        "summary": "Read one version.",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "versionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/apps/{appId}/config/diff": {
      "get": {
        "tags": ["Configuration"],
        "summary": "Compare two versions.",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiffResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfigChange": {
        "required": ["path", "kind", "from", "to"],
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "description": "RFC 6901 JSON Pointer to the value."
          },
          "kind": {
            "type": "string",
            "description": "One of `added`, `removed`, or `changed`."
          },
          "from": {
            "type": "string",
            "description": "The previous value, canonicalised, or null when added.",
            "nullable": true
          },
          "to": {
            "type": "string",
            "description": "The new value, canonicalised, or null when removed.",
            "nullable": true
          }
        },
        "description": "What changed at one location."
      },
...
      "DiagnosticResponse": {
        "required": ["code", "severity", "path", "message", "docsUrl"],
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable, documented, searchable code."
          },
          "severity": {
            "type": "string",
            "description": "Whether it blocks a save."
          },
          "path": {
            "type": "string",
            "description": "RFC 6901 JSON Pointer to the offending value."
          },
          "message": {
            "type": "string",
            "description": "User-facing text that names the fix."
          },
          "docsUrl": {
            "type": "string",
            "description": "Where to read more."
          }
        },
        "description": "One finding, as the API reports it."
      },
      "DiffResponse": {
        "required": ["from", "to", "changes"],
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "description": "The earlier version.",
            "format": "uuid"
          },
          "to": {
            "type": "string",
            "description": "The later version.",
            "format": "uuid"
          },
          "changes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConfigChange"
            },
            "description": "What differs, ordered by path."
          }
        },
        "description": "What a diff returned."
      },
...
      "JsonObject": {
        "type": "object"
      },
...
      "PageOfVersionSummary": {
        "required": ["items", "nextCursor"],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VersionSummary"
            },
            "description": "The page."
          },
          "nextCursor": {
            "type": "string",
            "description": "Pass back as `cursor` for the following page, or null at the end.",
            "nullable": true
          }
        },
        "description": "A page of results."
      },
...
      "SaveResponse": {
        "required": ["version", "created", "warnings", "info"],
        "type": "object",
        "properties": {
          "version": {
            "$ref": "#/components/schemas/VersionSummary"
          },
          "created": {
            "type": "boolean",
            "description": "False when an identical version already existed."
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticResponse"
            },
            "description": "Findings that did not block the save."
          },
          "info": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticResponse"
            },
            "description": "Hints."
          }
        },
        "description": "What a save returned."
      },
...
      "ValidationResponse": {
        "required": ["valid", "errors", "warnings", "info"],
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean",
            "description": "True when there are no errors."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticResponse"
            },
            "description": "Findings that block a save."
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticResponse"
            },
            "description": "Findings allowed through."
          },
          "info": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticResponse"
            },
            "description": "Hints."
          }
        },
        "description": "What a validate returned."
      },
      "VersionResponse": {
        "required": ["version", "config"],
        "type": "object",
        "properties": {
          "version": {
            "$ref": "#/components/schemas/VersionSummary"
          },
          "config": {
            "$ref": "#/components/schemas/JsonObject"
          }
        },
        "description": "A configuration version, body included."
      },
      "VersionSummary": {
        "required": ["id", "schemaVersion", "codeKey", "assetKey", "contentKey", "createdBy", "createdAt", "message"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identifier.",
            "format": "uuid"
          },
          "schemaVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "description": "Schema version the body conforms to.",
            "format": "int32"
          },
          "codeKey": {
            "type": "string",
            "description": "Cache key covering everything that forces a native recompile."
          },
          "assetKey": {
            "type": "string",
            "description": "Cache key covering everything that needs only a resource repackage."
          },
          "contentKey": {
            "type": "string",
            "description": "Cache key covering everything that needs only a config patch."
          },
          "createdBy": {
            "type": "string",
            "description": "Who saved it.",
            "format": "uuid",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "When.",
            "format": "date-time"
          },
          "message": {
            "type": "string",
            "description": "Optional note.",
            "nullable": true
          }
        },
        "description": "A configuration version's identity, without its body."
      },
...
    }
  },
...
}